# [onert] Apply ExternalTensor to cpu backend