Apply ExternalTensor to cpu backend. Now cpu backend's const tensors are
ExternalTensor so that are not memory-allocated.

ONE-DCO-1.0-Signed-off-by: Yongseop Kim <yons.kim@samsung.com>