@YongseopKim 2020-07-27
before

```
odroid@odroid:/home/dragon/Works/github/ONE$ backends="cpu" ./Product/out/bin/nnpackage_run -m1 -w5 -r1 ./benchmark_nnpkg_models/inception_v3/
Package Filename ./benchmark_nnpkg_models/inception_v3/
===================================
MODEL_LOAD   takes 35.898 ms
PREPARE      takes 2662.650 ms
EXECUTE      takes 1233.100 ms
- MEAN     :  1233.100 ms
- MAX      :  1233.100 ms
- MIN      :  1233.100 ms
- GEOMEAN  :  1233.100 ms
===================================
RSS
- MODEL_LOAD   takes 13964 kb
- PREPARE      takes 223048 kb
- EXECUTE      takes 141356 kb
- PEAK         takes 223048 kb
===================================
HWM
- MODEL_LOAD   takes 13964 kb
- PREPARE      takes 229472 kb
- EXECUTE      takes 229472 kb
- PEAK         takes 229472 kb
===================================
PSS
- MODEL_LOAD   takes 6142 kb
- PREPARE      takes 224776 kb
- EXECUTE      takes 139943 kb
- PEAK         takes 224776 kb
===================================
```

after

```
odroid@odroid:/home/dragon/Works/github/ONE$ backends="cpu" ./Product/out/bin/nnpackage_run -m1 -w5 -r1 ./benchmark_nnpkg_models/inception_v3/
Package Filename ./benchmark_nnpkg_models/inception_v3/
===================================
MODEL_LOAD   takes 42.920 ms
PREPARE      takes 2461.624 ms
EXECUTE      takes 1168.381 ms
- MEAN     :  1168.381 ms
- MAX      :  1168.381 ms
- MIN      :  1168.381 ms
- GEOMEAN  :  1168.381 ms
===================================
RSS
- MODEL_LOAD   takes 10864 kb
- PREPARE      takes 135840 kb
- EXECUTE      takes 140814 kb
- PEAK         takes 140814 kb
===================================
HWM
- MODEL_LOAD   takes 10864 kb
- PREPARE      takes 135840 kb
- EXECUTE      takes 140916 kb
- PEAK         takes 140916 kb
===================================
PSS
- MODEL_LOAD   takes 5998 kb
- PREPARE      takes 132858 kb
- EXECUTE      takes 138868 kb
- PEAK         takes 138868 kb
===================================
```